+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 2012-12-27
-Removed 'arc' related code.  The position of the meter in the new TT layout makes the arc almost un-noticable.
+- Stripped the original script down to the bare minimum.  No auto voting, no arc, no 'bot message'.
+- Added a bit of code for a menu item in the Turntable Settings menu.